requirements.txt
fastapi==0.110
uvicorn==0.29.0
colorama==0.4.6
google-cloud-storage==2.10.0
google-cloud==0.34.0
google-cloud-secret-manager==2.16.4
python-dotenv==1.0.1
autopep
black
pymongo
redis
openai==1.0.0